requirements.txt
djangorestframework
Django
joblib
scikit-learn
psycopg2
pandas 
numpy